requirements/prod.txt
# Everything needed in production

# Flask
Flask>=1.0.2
MarkupSafe
Werkzeug
Jinja2>=2.10
itsdangerous
click>=5.0
flask_moment>=0.5.1
Flask-API

# Database
Flask-SQLAlchemy
psycopg2
SQLAlchemy

# Migrations
Flask-Migrate

# Forms
Flask-WTF
WTForms

# Deployment
gunicorn>=19.1.1

# Assets
Flask-Assets
cssmin>=0.2.0
jsmin>=2.0.11

# File upload and images
flask-uploads>=0.2.1
Pillow>=4.1.0

# Auth
Flask-Security
Flask-Bcrypt
flask-kvsession
flask-admin

# Caching
Flask-Caching>=1.0.0

# Debug toolbar
Flask-DebugToolbar

# Crontab
Flask-APScheduler

# Payment system
btcpay-python

# Notifications
python-pushover
#icinga2api

# Other
names>=0.3.0
pymysql
requests
python-dateutil
oslo_concurrency
